Adapt Unit Tests
