+++ app/src/main/java/com/arygm/quickfix/model/profile/Profile.kt
@@ -10,7 +10,23 @@ open class Profile(
     val email: String,
     val birthDate: Timestamp,
     val location: GeoPoint? = null
-)
+) {
+  override fun equals(other: Any?): Boolean {
+    if (this === other) return true
+    if (other !is Profile) return false
+
+    return uid == other.uid &&
+        firstName == other.firstName &&
+        lastName == other.lastName &&
+        email == other.email &&
+        birthDate == other.birthDate &&
+        location == other.location
+  }
+
+  override fun hashCode(): Int {
+    return listOf(uid, firstName, lastName, email, birthDate, location).hashCode()
+  }
+}
 
 class UserProfile(
     val isWorker: Boolean = false,
@@ -20,7 +36,20 @@ class UserProfile(
     email: String,
     birthDate: Timestamp,
     location: GeoPoint? = null
-) : Profile(uid, firstName, lastName, email, birthDate, location)
+) : Profile(uid, firstName, lastName, email, birthDate, location) {
+
+  override fun equals(other: Any?): Boolean {
+    if (this === other) return true
+    if (other !is UserProfile) return false
+    if (!super.equals(other)) return false
+
+    return isWorker == other.isWorker
+  }
+
+  override fun hashCode(): Int {
+    return listOf(super.hashCode(), isWorker).hashCode()
+  }
+}
 
 class WorkerProfile(
     val fieldOfWork: String? = null,
@@ -32,7 +61,20 @@ class WorkerProfile(
     email: String,
     birthDate: Timestamp,
     location: GeoPoint? = null
-) : Profile(uid, firstName, lastName, email, birthDate, location)
+) : Profile(uid, firstName, lastName, email, birthDate, location) {
+
+  override fun equals(other: Any?): Boolean {
+    if (this === other) return true
+    if (other !is WorkerProfile) return false
+    if (!super.equals(other)) return false
+
+    return fieldOfWork == other.fieldOfWork && hourlyRate == other.hourlyRate
+  }
+
+  override fun hashCode(): Int {
+    return listOf(super.hashCode(), fieldOfWork, hourlyRate).hashCode()
+  }
+}
 
 sealed class WorkerCategory {
   sealed class ConstructionAndMaintenance : WorkerCategory() {
